requirements.txt
streamlit
langchain-google-genai
langchain-core
python-dotenv
pydantic
